The goals of this code are: 1. Read in/Parse an arbitrary number of Excel files (workbooks). 2. Merge the contents from each identically named worksheet in those workbooks. (This is known more generally as “group by”. We must group by worksheet name.) 3. Write a single workbook with multiple worksheets. Each output worksheet contains the merged contribution from the worksheets from each workbook that share a common name. The worksheets in the final workbook will be the union of worksheets from all workbooks parsed in step #1.
The main() function is where the files are parsed in. In the excel_parser function, data from identically named worksheets are grouped together and analyzed for merge conflicts. In this code, if a conflict is detected, it prompts user input to determine the right value.

As for the argparse functions, it has the -o, --output function which allows the user to specify the output file. It also has the -i, --input function to specify the type of the file ('xlsx', 'txt', etc.) In the future, it will also have a -m, --merge-default function to specify a default resolution to a conflict, as well as a --output-format function to specify the type of the output file.