The goals of this code are: 1. Read in/Parse an arbitrary number of Excel files (workbooks). 2. Merge the contents from each identically named worksheet in those workbooks. (This is known more generally as “group by”. We must group by worksheet name.) 3. Write a single workbook with multiple worksheets. Each output worksheet contains the merged contribution from the worksheets from each workbook that share a common name. The worksheets in the final workbook will be the union of worksheets from all workbooks parsed in step #1.
The main() function is where the files are parsed in. In the write_excel function, data from identically named worksheets are grouped together and analyzed for merge conflicts.

All of the argparse functions are written, including '-i' to specify input file type, '-o' to specify the output file, '--interactive' to let the user chose the correct value in case of a merge conflict, '--mergedefault' to specify the default action in case a merge conflict arises, and '-q' to quietly abort the merge if there is a conflict.