Citrine ingester (#18)
* Prototype Citrine ingester complete, but in the interim,
changes were made to the underlying CLI arguments. This
commits the current state of the ingester, allowing the
updated CLI to be pulled and merged.

* Prototype Citrine ingester complete, but in the interim,
changes were made to the underlying CLI arguments. This
commits the current state of the ingester, allowing the
updated CLI to be pulled and merged.